# ApiGateway/app/openapi.yaml
openapi: 3.0.0
info:
  title: Chat Application API Gateway (SOA/CQRS)
  version: 1.0.0
  description: |
    API Gateway serving Chat Command, Chat Query, and Customer Management services.
    
    **NOTE:**
    1. Public Endpoints: `/auth/login`, `/auth/logout`, `/customer/create/account`, and `/customer/forgetPass/**`.
    2. Authenticated Endpoints: All other `/customer`, `/command`, and `/query` endpoints require the `jwt_token` cookie.
    3. **Internal Endpoints (Service-to-Service)**: Endpoints starting with `/internal/**` are for internal communication between services only and should not be called directly from client/frontend.
servers:
  - url: http://localhost:8080
    description: Nginx Gateway (Local Development)
security:
  - cookieAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: jwt_token
      description: JWT Token set after successful login.
  schemas:
    GenericResponse:
      type: object
      properties:
        success:
          type: boolean
          description: True if request is successful, False if there is a logical or business error.
        message:
          type: string
          description: Detailed notification about the result.
        data:
          type: object
          nullable: true
          description: Returned data (depends on the endpoint).
      required:
        - success
        - message
    
    # --- Auth & Customer Schemas ---
...
    Customer:
      type: object
      properties:
        customerId:
          type: integer
          format: int64
          description: Unique ID of the customer.
          example: 1
        username:
          type: string
          example: user1
        fullName:
          type: string
          example: Vũ Thanh Hải
        email:
          type: string
          example: vuhaipro2707@gmail.com
        phoneNumber:
          type: string
          example: '0901234567'
        avatarColor:
          type: string
          example: '#4CAF50'
    UpdateCustomerInfoRequest:
      type: object
      properties:
        fullName:
          type: string
          example: Vũ Thanh Hải (Mới)
        email:
          type: string
          example: vuhaipro2707_new@gmail.com
        phoneNumber:
          type: string
          example: '0901234569'
        avatarColor:
          type: string
          example: '#FF5733'
      required:
        - fullName
        - email
        - avatarColor
    CreateCustomerRequest:
      type: object
      properties:
        username:
          type: string
          example: newuser
        password:
          type: string
          example: newpass123
        fullName:
          type: string
          example: User Mới
        email:
          type: string
          example: newuser@example.com
        phoneNumber:
          type: string
          example: '0912345678'
        avatarColor:
          type: string
          example: '#33AACC'
      required:
        - username
        - password
        - fullName
        - email
    ForgetPasswordInitiateRequest:
      type: object
      properties:
        email:
          type: string
          example: user1@example.com
      required:
        - email
    ForgetPasswordConfirmRequest:
      type: object
      properties:
        email:
          type: string
          example: user1@example.com
        otpCode:
          type: string
          example: '123456'
      required:
        - email
        - otpCode
    PasswordResetRequest:
      type: object
      properties:
        email:
          type: string
          example: user1@example.com
        newPassword:
          type: string
          example: NewPassword123
      required:
        - email
        - newPassword
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
          example: oldpassword
        newPassword:
          type: string
          example: newpassword
      required:
        - currentPassword
        - newPassword
        
    # --- OTP & Mail Service Schemas (Internal) ---
...
paths:
...
  /customer/create/account:
    post:
      tags:
        - Customer Management Service (Public)
      summary: Create a new customer account
      operationId: createCustomerAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
      responses:
        '200':
          description: Account created successfully.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/GenericResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Customer'
        '400':
          description: Invalid input data (e.g., username/email already exists).
  /customer/forgetPass/initiate:
    post:
      tags:
        - Customer Management Service (Public)
      summary: Initiate password reset by sending OTP to email
      operationId: forgetPassInitiate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgetPasswordInitiateRequest'
      responses:
        '200':
          description: OTP sent to email.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
              example:
                success: true
                message: OTP sent to email.
                data: null
        '404':
          description: Email not found.
  /customer/forgetPass/resend:
    post:
      tags:
        - Customer Management Service (Public)
      summary: Resend password reset OTP to email
      operationId: forgetPassResend
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgetPasswordInitiateRequest'
      responses:
        '200':
          description: OTP resent successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
              example:
                success: true
                message: OTP resent to email.
                data: null
        '404':
          description: Email not found.
  /customer/forgetPass/confirm:
    post:
      tags:
        - Customer Management Service (Public)
      summary: Confirm OTP for password reset
      operationId: forgetPassConfirm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgetPasswordConfirmRequest'
      responses:
        '200':
          description: OTP confirmed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
              example:
                success: true
                message: OTP confirmed. You can now reset your password.
                data: null
        '400':
          description: Invalid or expired OTP or session.
  /customer/forgetPass/reset:
    post:
      tags:
        - Customer Management Service (Public)
      summary: Reset password using confirmed session
      operationId: forgetPassReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '200':
          description: Password reset successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
              example:
                success: true
                message: Password reset successfully.
                data: null
        '400':
          description: OTP confirmation is required or invalid input.
        '404':
          description: Email not found.
  /customer/changePass:
    post:
      tags:
        - Customer Management Service (Auth Required)
      summary: Change password for authenticated user
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
              example:
                success: true
                message: Password changed successfully.
                data: null
        '400':
          description: Invalid input (e.g., missing fields, new password equals old password).
        '401':
          description: Invalid current password.
  /customer/info:
    get:
      tags:
        - Customer Management Service (Auth Required)
      summary: Get current user's personal info
      operationId: getCurrentUserInfo
      responses:
        '200':
          description: Successfully retrieved info.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/GenericResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Customer'
        '401':
          description: Unauthorized (JWT Cookie missing or invalid/expired).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
              example:
                success: false
                message: Invalid authentication credentials
                data: null
    post:
      tags:
        - Customer Management Service (Auth Required)
      summary: Update current user's personal info
      operationId: updateCurrentUserInfo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomerInfoRequest'
      responses:
        '200':
          description: Update successful.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/GenericResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Customer'
        '400':
          description: Invalid input data (e.g., missing fields, email/phone already exists).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
              examples:
                missingFields:
                  value:
                    success: false
                    message: Full name, email, and avatar color are required.
                emailInUse:
                  value:
                    success: false
                    message: Email already in use.
        '401':
          description: Unauthorized.
  /customer/info/customerId/{customerId}:
    get:
      tags:
        - Customer Management Service (Auth Required)
      summary: Get detailed Customer info by ID
      operationId: getCustomerInfoById
      parameters:
        - in: path
          name: customerId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved info.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/GenericResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Customer'
        '404':
          description: Customer does not exist.
        '401':
          description: Unauthorized.
  /customer/fullName/customerId/{customerId}:
    get:
      tags:
        - Customer Management Service (Auth Required)
      summary: Get Customer's Full Name by ID (Used by Chat Command Service)
      operationId: getCustomerFullNameById
      parameters:
        - in: path
          name: customerId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved name.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/GenericResponse'
                  - type: object
                    properties:
                      data:
                        type: string
                        example: Vũ Thanh Hải
        '404':
          description: Customer does not exist.
  /customer/info/phoneNumber/{phoneNumber}:
    get:
      tags:
        - Customer Management Service (Auth Required)
      summary: Search Customer by phone number (Contains)
      operationId: searchCustomerByPhoneNumber
      parameters:
        - in: path
          name: phoneNumber
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of found Customers (can be empty).
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/GenericResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Customer'
        '400':
          description: Phone number missing.
        '401':
          description: Unauthorized.
  /customer/info/fullName/{fullName}:
    get:
      tags:
        - Customer Management Service (Auth Required)
      summary: Search Customer by Full Name (Contains)
      operationId: searchCustomerByFullName
      parameters:
        - in: path
          name: fullName
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of found Customers (can be empty).
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/GenericResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Customer'
        '400':
          description: Full name missing.
        '401':
          description: Unauthorized.
          
  # =================================================================
  # CHAT COMMAND ENDPOINTS (Auth Required, Writes to PostgreSQL/Kafka)
  # =================================================================
...