# customer-management-service/src/main/java/com/example/customer_management_service/dto/ForgetPasswordInitiateRequest.java
package com.example.customer_management_service.dto;

import lombok.Data;
import lombok.AllArgsConstructor;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class ForgetPasswordInitiateRequest {
    private String email;
}